+++ server/schema/getchangedzones.sql
@@ -1,6 +1,6 @@
 CREATE OR REPLACE FUNCTION GetChangedZones(
 	nameservername varchar,
-	out change_id int,
+	out change_id bigint,
 	out change_name varchar,
 	out change_changetime int
 ) RETURNS SETOF record AS $$
